Changed the diff controller to have a 32 bit int instead of 16, which is I suspect is a reason for the I'm getting with the wheel, and the reason the robot hit a parked car

+++ src/differential-drive/src/pid_velocity.cpp
@@ -20,6 +20,7 @@
 
 #include "ros/ros.h"
 #include "std_msgs/Int16.h"
+#include "std_msgs/Int32.h"
 #include "std_msgs/Float32.h"
 
 #include <sstream>
@@ -85,7 +86,7 @@ private:
 
 public:
 
-    void wheelCallback(std_msgs::Int16 message) {
+    void wheelCallback(std_msgs::Int32 message) {
         int enc = (int)message.data;
         if (enc < encoder_low_wrap and prev_encoder > encoder_high_wrap) {
             wheel_mult = wheel_mult + 1;
